requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
python-dotenv==0.19.0
pydantic==1.8.2
nltk==3.4.5
pdfminer.six==20181108
docx2txt==0.7
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sortedcontainers==2.1.0
urllib3==1.25.3
email-validator==1.1.3
alembic==1.7.1